Fixed scaling issues

+++ PdfPixel.PdfPanel.Web/WorkerInterface/UpdateContentRequest.cs
@@ -4,5 +4,5 @@ public class UpdateContentRequest : ContentRequest
 {
     public int PageNumber { get; set; }
 
-    public int Scale { get; set; }
+    public float Scale { get; set; }
 }
